Read JSON config from console
run function can not read JSON file, which will supercede our previous use of sourcing an R file. These changes shouldn't break shiny, but needs confirming. Still possible to read in a config.R, but a warning is created stating that this is depreciated. I've committed versions of both configs used in the 10-year FCF paper for reference.