initial commit
working to simplify constraints to only 3 fields: (1) fixed vs variable toggle, (2) constraint field, (3) constraint value. Currently, (3) is divided into values for fixed (e.g., 2000) and variable (e.g., 20%). Constraints are applied to the constraint field, aggregated to the the proj id, and are applied either as a fixed value (e.g., 2000 of constraint field @ project scale) or as a percent value (e.g., 20% of constraint field aggregated to project). This should be much more generalizable (i.e., not just area) and have all the functionality as before, but with significantly fewer parameters. The changes are significant enough to warrant review, however.